applying some of jims suggestions
