+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene92/Lucene92HnswVectorsWriter.java
@@ -26,6 +26,7 @@ import java.nio.ByteOrder;
 import java.util.Arrays;
 import org.apache.lucene.codecs.BufferingKnnVectorsWriter;
 import org.apache.lucene.codecs.CodecUtil;
+import org.apache.lucene.codecs.hnsw.DefaultFlatVectorScorer;
 import org.apache.lucene.codecs.lucene90.IndexedDISI;
 import org.apache.lucene.index.ByteVectorValues;
 import org.apache.lucene.index.DocsWithFieldSet;
@@ -275,10 +276,10 @@ public final class Lucene92HnswVectorsWriter extends BufferingKnnVectorsWriter {
   private OnHeapHnswGraph writeGraph(
       RandomAccessVectorValues.Floats vectorValues, VectorSimilarityFunction similarityFunction)
       throws IOException {
-
+    DefaultFlatVectorScorer defaultFlatVectorScorer = new DefaultFlatVectorScorer();
     // build graph
     RandomVectorScorerSupplier scorerSupplier =
-        RandomVectorScorerSupplier.createFloats(vectorValues, similarityFunction);
+        defaultFlatVectorScorer.getRandomVectorScorerSupplier(similarityFunction, vectorValues);
     HnswGraphBuilder hnswGraphBuilder =
         HnswGraphBuilder.create(scorerSupplier, M, beamWidth, HnswGraphBuilder.randSeed);
     hnswGraphBuilder.setInfoStream(segmentWriteState.infoStream);
