+++ lucene/core/src/java/org/apache/lucene/util/hnsw/RandomVectorScorer.java
@@ -18,7 +18,6 @@
 package org.apache.lucene.util.hnsw;
 
 import java.io.IOException;
-import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.util.Bits;
 
 /** A {@link RandomVectorScorer} for scoring random nodes in batches against an abstract query. */
@@ -56,58 +55,6 @@ public interface RandomVectorScorer {
     return acceptDocs;
   }
 
-  /**
-   * Creates a default scorer for float vectors.
-   *
-   * <p>WARNING: The {@link RandomAccessVectorValues} given can contain stateful buffers. Avoid
-   * using it after calling this function. If you plan to use it again outside the returned {@link
-   * RandomVectorScorer}, think about passing a copied version ({@link
-   * RandomAccessVectorValues#copy}).
-   *
-   * @param vectors the underlying storage for vectors
-   * @param similarityFunction the similarity function to score vectors
-   * @param query the actual query
-   */
-  static FloatVectorScorer createFloats(
-      final RandomAccessVectorValues.Floats vectors,
-      final VectorSimilarityFunction similarityFunction,
-      final float[] query) {
-    if (query.length != vectors.dimension()) {
-      throw new IllegalArgumentException(
-          "vector query dimension: "
-              + query.length
-              + " differs from field dimension: "
-              + vectors.dimension());
-    }
-    return new FloatVectorScorer(vectors, query, similarityFunction);
-  }
-
-  /**
-   * Creates a default scorer for byte vectors.
-   *
-   * <p>WARNING: The {@link RandomAccessVectorValues} given can contain stateful buffers. Avoid
-   * using it after calling this function. If you plan to use it again outside the returned {@link
-   * RandomVectorScorer}, think about passing a copied version ({@link
-   * RandomAccessVectorValues#copy}).
-   *
-   * @param vectors the underlying storage for vectors
-   * @param similarityFunction the similarity function to use to score vectors
-   * @param query the actual query
-   */
-  static ByteVectorScorer createBytes(
-      final RandomAccessVectorValues.Bytes vectors,
-      final VectorSimilarityFunction similarityFunction,
-      final byte[] query) {
-    if (query.length != vectors.dimension()) {
-      throw new IllegalArgumentException(
-          "vector query dimension: "
-              + query.length
-              + " differs from field dimension: "
-              + vectors.dimension());
-    }
-    return new ByteVectorScorer(vectors, query, similarityFunction);
-  }
-
   /** Creates a default scorer for random access vectors. */
   abstract class AbstractRandomVectorScorer implements RandomVectorScorer {
     private final RandomAccessVectorValues values;
@@ -134,50 +81,6 @@ public interface RandomVectorScorer {
     @Override
     public Bits getAcceptOrds(Bits acceptDocs) {
       return values.getAcceptOrds(acceptDocs);
-    }
-  }
-
-  /** A {@link RandomVectorScorer} for float vectors. */
-  class FloatVectorScorer extends AbstractRandomVectorScorer {
-    private final RandomAccessVectorValues.Floats values;
-    private final float[] query;
-    private final VectorSimilarityFunction similarityFunction;
-
-    public FloatVectorScorer(
-        RandomAccessVectorValues.Floats values,
-        float[] query,
-        VectorSimilarityFunction similarityFunction) {
-      super(values);
-      this.values = values;
-      this.query = query;
-      this.similarityFunction = similarityFunction;
-    }
-
-    @Override
-    public float score(int node) throws IOException {
-      return similarityFunction.compare(query, values.vectorValue(node));
-    }
-  }
-
-  /** A {@link RandomVectorScorer} for byte vectors. */
-  class ByteVectorScorer extends AbstractRandomVectorScorer {
-    private final RandomAccessVectorValues.Bytes values;
-    private final byte[] query;
-    private final VectorSimilarityFunction similarityFunction;
-
-    public ByteVectorScorer(
-        RandomAccessVectorValues.Bytes values,
-        byte[] query,
-        VectorSimilarityFunction similarityFunction) {
-      super(values);
-      this.values = values;
-      this.query = query;
-      this.similarityFunction = similarityFunction;
-    }
-
-    @Override
-    public float score(int node) throws IOException {
-      return similarityFunction.compare(query, values.vectorValue(node));
     }
   }
 }
